GCD: re-org location (only, not functionality)
Uncontroversial, I think:

- gcd centrally
- Stolzenburg in consonance
- only those specific to rhythm/metre stay where they are
